Change floats to doubles
ACCOUNTING
Add method getContactByContactNumber

LineItem Quantity change from float to double
LineItem UnitAmount change from float to double
BankTransfer Amound change from float to double
Item QuantityOnHand change from float to double
PurchaseOrder UnitPrice change from float to double

ASSETS, PAYROLL AU and PAYROLL UK
All numbers with format float changed to double

+++ src/test/java/com/xero/api/client/PayrollAuApiTimesheetTest.java
@@ -70,21 +70,21 @@ public class PayrollAuApiTimesheetTest {
     public void getTimesheetsTest() throws IOException {
         System.out.println("@Test - getTimesheetsTest");
      
-        List<Float> numOfUnitList = new ArrayList<>();
-        numOfUnitList.add(4.0f);
-        numOfUnitList.add(4.0f);
-        numOfUnitList.add(4.0f);
-        numOfUnitList.add(4.0f);
-        numOfUnitList.add(4.0f);
-        numOfUnitList.add(4.0f);
-        numOfUnitList.add(0.0f);
+        List<Double> numOfUnitList = new ArrayList<>();
+        numOfUnitList.add(4.0);
+        numOfUnitList.add(4.0);
+        numOfUnitList.add(4.0);
+        numOfUnitList.add(4.0);
+        numOfUnitList.add(4.0);
+        numOfUnitList.add(4.0);
+        numOfUnitList.add(0.0);
         Timesheets response = payrollAuApi.getTimesheets(accessToken, xeroTenantId, null, null, null, null);
         
         assertThat(response.getTimesheets().get(0).getEmployeeID(), is(equalTo(UUID.fromString("b34e89ff-770d-4099-b7e5-f968767118bc"))));       
         assertThat(response.getTimesheets().get(0).getStartDateAsDate() , is(equalTo(LocalDate.of(2019,01,18))));
         assertThat(response.getTimesheets().get(0).getEndDateAsDate() , is(equalTo(LocalDate.of(2019,01,24))));  
         assertThat(response.getTimesheets().get(0).getStatus(), is(equalTo(com.xero.models.payrollau.TimesheetStatus.APPROVED)));
-        assertThat(response.getTimesheets().get(0).getHours(), is(equalTo(24.0f)));        
+        assertThat(response.getTimesheets().get(0).getHours(), is(equalTo(24.0)));        
         assertThat(response.getTimesheets().get(0).getTimesheetID(), is(equalTo(UUID.fromString("863bbd31-0447-4419-80d5-d733d5e723ba"))));       
         assertThat(response.getTimesheets().get(0).getTimesheetLines().get(0).getEarningsRateID(), is(equalTo(UUID.fromString("ab874dfb-ab09-4c91-954e-43acf6fc23b4"))));       
         assertThat(response.getTimesheets().get(0).getTimesheetLines().get(0).getNumberOfUnits(), is(equalTo(numOfUnitList)));       
@@ -98,14 +98,14 @@ public class PayrollAuApiTimesheetTest {
         System.out.println("@Test - getTimesheetTest");
         UUID timesheetId = UUID.fromString("cdfb8371-0b21-4b8a-8903-1024df6c391e");
        
-        List<Float> numOfUnitList = new ArrayList<>();
-        numOfUnitList.add(3.0f);
-        numOfUnitList.add(3.0f);
-        numOfUnitList.add(3.0f);
-        numOfUnitList.add(3.0f);
-        numOfUnitList.add(0.0f);
-        numOfUnitList.add(3.0f);
-        numOfUnitList.add(0.0f);
+        List<Double> numOfUnitList = new ArrayList<>();
+        numOfUnitList.add(3.0);
+        numOfUnitList.add(3.0);
+        numOfUnitList.add(3.0);
+        numOfUnitList.add(3.0);
+        numOfUnitList.add(0.0);
+        numOfUnitList.add(3.0);
+        numOfUnitList.add(0.0);
         
         TimesheetObject response = payrollAuApi.getTimesheet(accessToken, xeroTenantId, timesheetId);
         
@@ -113,7 +113,7 @@ public class PayrollAuApiTimesheetTest {
         assertThat(response.getTimesheet().getStartDateAsDate() , is(equalTo(LocalDate.of(2019,01,11))));
         assertThat(response.getTimesheet().getEndDateAsDate() , is(equalTo(LocalDate.of(2019,01,17))));  
         assertThat(response.getTimesheet().getStatus(), is(equalTo(com.xero.models.payrollau.TimesheetStatus.APPROVED)));
-        assertThat(response.getTimesheet().getHours(), is(equalTo(15.0f)));
+        assertThat(response.getTimesheet().getHours(), is(equalTo(15.0)));
         assertThat(response.getTimesheet().getTimesheetID(), is(equalTo(UUID.fromString("df954ca3-3a70-47e9-9a3e-80711e7c5f90"))));       
         assertThat(response.getTimesheet().getTimesheetLines().get(0).getEarningsRateID(), is(equalTo(UUID.fromString("ab874dfb-ab09-4c91-954e-43acf6fc23b4"))));       
         assertThat(response.getTimesheet().getTimesheetLines().get(0).getNumberOfUnits(), is(equalTo(numOfUnitList)));       
@@ -128,14 +128,14 @@ public class PayrollAuApiTimesheetTest {
         System.out.println("@Test - createTimesheetTest");
         UUID timesheetId = UUID.fromString("cdfb8371-0b21-4b8a-8903-1024df6c391e");
         List<Timesheet> timesheets = new ArrayList<>();
-        List<Float> numOfUnitList = new ArrayList<>();
-        numOfUnitList.add(2.0f);
-        numOfUnitList.add(10.0f);
-        numOfUnitList.add(0.0f);
-        numOfUnitList.add(0.0f);
-        numOfUnitList.add(5.0f);
-        numOfUnitList.add(0.0f);
-        numOfUnitList.add(5.0f);
+        List<Double> numOfUnitList = new ArrayList<>();
+        numOfUnitList.add(2.0);
+        numOfUnitList.add(10.0);
+        numOfUnitList.add(0.0);
+        numOfUnitList.add(0.0);
+        numOfUnitList.add(5.0);
+        numOfUnitList.add(0.0);
+        numOfUnitList.add(5.0);
         
         Timesheets response = payrollAuApi.createTimesheet(accessToken, xeroTenantId, timesheets);
         
@@ -143,7 +143,7 @@ public class PayrollAuApiTimesheetTest {
         assertThat(response.getTimesheets().get(0).getStartDateAsDate() , is(equalTo(LocalDate.of(2019,11,8))));
         assertThat(response.getTimesheets().get(0).getEndDateAsDate() , is(equalTo(LocalDate.of(2019,11,14))));  
         assertThat(response.getTimesheets().get(0).getStatus(), is(equalTo(com.xero.models.payrollau.TimesheetStatus.DRAFT)));
-        assertThat(response.getTimesheets().get(0).getHours(), is(equalTo(22.0f)));        
+        assertThat(response.getTimesheets().get(0).getHours(), is(equalTo(22.0)));        
         assertThat(response.getTimesheets().get(0).getTimesheetID(), is(equalTo(UUID.fromString("a7eb0a79-8511-4ee7-b473-3a25f28abcb9"))));       
         assertThat(response.getTimesheets().get(0).getTimesheetLines().get(0).getEarningsRateID(), is(equalTo(UUID.fromString("ab874dfb-ab09-4c91-954e-43acf6fc23b4"))));       
         assertThat(response.getTimesheets().get(0).getTimesheetLines().get(0).getTrackingItemID(), is(equalTo(UUID.fromString("af5e9ce2-2349-4136-be99-3561b189f473"))));       
@@ -158,14 +158,14 @@ public class PayrollAuApiTimesheetTest {
         System.out.println("@Test - updateTimesheetTest");
         UUID timesheetId = UUID.fromString("cdfb8371-0b21-4b8a-8903-1024df6c391e");
         List<Timesheet> timesheets = new ArrayList<>();
-        List<Float> numOfUnitList = new ArrayList<>();
-        numOfUnitList.add(2.0f);
-        numOfUnitList.add(10.0f);
-        numOfUnitList.add(0.0f);
-        numOfUnitList.add(0.0f);
-        numOfUnitList.add(5.0f);
-        numOfUnitList.add(0.0f);
-        numOfUnitList.add(5.0f);
+        List<Double> numOfUnitList = new ArrayList<>();
+        numOfUnitList.add(2.0);
+        numOfUnitList.add(10.0);
+        numOfUnitList.add(0.0);
+        numOfUnitList.add(0.0);
+        numOfUnitList.add(5.0);
+        numOfUnitList.add(0.0);
+        numOfUnitList.add(5.0);
         
         Timesheets response = payrollAuApi.updateTimesheet(accessToken, xeroTenantId, timesheetId, timesheets);
         
@@ -173,7 +173,7 @@ public class PayrollAuApiTimesheetTest {
         assertThat(response.getTimesheets().get(0).getStartDateAsDate() , is(equalTo(LocalDate.of(2019,11,8))));
         assertThat(response.getTimesheets().get(0).getEndDateAsDate() , is(equalTo(LocalDate.of(2019,11,14))));  
         assertThat(response.getTimesheets().get(0).getStatus(), is(equalTo(com.xero.models.payrollau.TimesheetStatus.APPROVED)));
-        assertThat(response.getTimesheets().get(0).getHours(), is(equalTo(22.0f)));        
+        assertThat(response.getTimesheets().get(0).getHours(), is(equalTo(22.0)));        
         assertThat(response.getTimesheets().get(0).getTimesheetID(), is(equalTo(UUID.fromString("a7eb0a79-8511-4ee7-b473-3a25f28abcb9"))));       
         assertThat(response.getTimesheets().get(0).getTimesheetLines().get(0).getEarningsRateID(), is(equalTo(UUID.fromString("ab874dfb-ab09-4c91-954e-43acf6fc23b4"))));       
         assertThat(response.getTimesheets().get(0).getTimesheetLines().get(0).getTrackingItemID(), is(equalTo(UUID.fromString("af5e9ce2-2349-4136-be99-3561b189f473"))));
